+++ platform-operator/controllers/verrazzano/component/authproxy/testdata/affinityOverrideValues.yaml
@@ -11,6 +11,10 @@ config:
   dnsSuffix: 11.22.33.44.nip.io
   envName: default
 
+dns:
+  wildcard:
+    domain: nip.io
+
 affinity: |
   nodeAffinity:
     requiredDuringSchedulingIgnoredDuringExecution:
