+++ platform-operator/controllers/verrazzano/component/authproxy/testdata/enabledOverrideValues.yaml
@@ -10,3 +10,7 @@ proxy:
 config:
   dnsSuffix: 11.22.33.44.nip.io
   envName: default
+
+dns:
+  wildcard:
+    domain: nip.io
