+++ platform-operator/controllers/verrazzano/component/authproxy/testdata/replicasOverrideValues.yaml
@@ -11,3 +11,7 @@ proxy:
 config:
   dnsSuffix: 11.22.33.44.nip.io
   envName: default
+
+dns:
+  wildcard:
+    domain: nip.io
