+++ platform-operator/controllers/verrazzano/component/verrazzano/verrazzano.go
@@ -160,13 +160,6 @@ func generateOverridesFile(ctx spi.ComponentContext, overrides *verrazzanoValues
 
 func appendVerrazzanoValues(ctx spi.ComponentContext, overrides *verrazzanoValues) error {
 	effectiveCR := ctx.EffectiveCR()
-	if isWildcardDNS, domain := getWildcardDNS(&effectiveCR.Spec); isWildcardDNS {
-		overrides.DNS = &dnsValues{
-			Wildcard: &wildcardDNSSettings{
-				Domain: domain,
-			},
-		}
-	}
 
 	dnsSuffix, err := vzconfig.GetDNSSuffix(ctx.Client(), effectiveCR)
 	if err != nil {
@@ -352,13 +345,6 @@ func appendFluentdOverrides(effectiveCR *vzapi.Verrazzano, overrides *verrazzano
 
 func isVMOEnabled(vz *vzapi.Verrazzano) bool {
 	return vzconfig.IsPrometheusEnabled(vz) || vzconfig.IsKibanaEnabled(vz) || vzconfig.IsElasticsearchEnabled(vz) || vzconfig.IsGrafanaEnabled(vz)
-}
-
-func getWildcardDNS(vz *vzapi.VerrazzanoSpec) (bool, string) {
-	if vz.Components.DNS != nil && vz.Components.DNS.Wildcard != nil {
-		return true, vz.Components.DNS.Wildcard.Domain
-	}
-	return false, ""
 }
 
 // This function is used to fixup the fluentd daemonset on a managed cluster so that helm upgrade of Verrazzano does
