+++ platform-operator/helm_config/charts/verrazzano-authproxy/values.yaml
@@ -42,3 +42,7 @@ affinity: |
 config:
   envName:
   dnsSuffix:
+
+dns:
+  wildcard:
+    domain:
