+++ platform-operator/helm_config/charts/verrazzano/values.yaml
@@ -133,7 +133,3 @@ kubernetes:
 
 externaldns:
   enabled: false
-
-dns:
-  wildcard:
-    domain:
